agente/requirements.txt
# ============================================================
#  ARIA - Dependências
# ============================================================

requests
pillow
opencv-python
pyautogui
pyperclip
pyttsx3

# Reconhecimento de voz:
SpeechRecognition
# Fallback de captura (quando PyAudio não está disponível)
sounddevice
# pyaudio → se pip install pyaudio falhar no Windows, use:
#   pip install pipwin
#   pipwin install pyaudio
pyaudio

# Download de videos por link
yt-dlp

# Bot Telegram
python-telegram-bot
